Move sniffer test to a separate file
